+++ packages/connectkit/src/components/ConnectKit.tsx
@@ -21,6 +21,7 @@ import { ThemeProvider } from 'styled-components';
 import { useThemeFont } from '../hooks/useGoogleFont';
 import { useAccount, useNetwork } from 'wagmi';
 import { SIWEContext } from './Standard/SIWE/SIWEContext';
+import { getGlobalChains } from '../defaultClient';
 
 export const routes = {
   ONBOARDING: 'onboarding',
@@ -117,7 +118,7 @@ export const ConnectKitProvider: React.FC<ConnectKitProviderProps> = ({
     disclaimer: null,
     bufferPolyfill: true,
     customAvatar: undefined,
-    initialChainId: undefined,
+    initialChainId: getGlobalChains()[0]?.id,
     ethereumOnboardingUrl: undefined,
     walletOnboardingUrl: undefined,
   };